% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_dataset_name.R
\name{check_dataset_name}
\alias{check_dataset_name}
\title{Throws an error if a dataset name is invalid}
\usage{
check_dataset_name(dataset_name)
}
\arguments{
\item{dataset_name}{a resource ID}
}
\description{
Used to attempt to validate a dataset name before submitting it to the API
}
