% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource.R
\name{ds_search_url}
\alias{ds_search_url}
\title{Creates the URL for the datastore search end-point}
\usage{
ds_search_url()
}
\value{
a url
}
\description{
Creates the URL for the datastore search end-point
}
