% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dump_download.R
\name{dump_download}
\alias{dump_download}
\title{Use datastore dump to download whole files}
\usage{
dump_download(id)
}
\arguments{
\item{id}{a resource ID}
}
\value{
dataframe containing resource records
}
\description{
Use datastore dump to download whole files
}
