% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/error_check.R
\name{error_check}
\alias{error_check}
\title{Throws errors found in the content of an \code{httr::GET} request}
\usage{
error_check(content)
}
\arguments{
\item{content}{object produced by \code{httr::content}}
}
\description{
Throws errors found in the content of an \code{httr::GET} request
}
