% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_all_resources.R
\name{get_all_resources}
\alias{get_all_resources}
\title{Provides an overview of all resources available from <opendata.nhs.scot>.}
\usage{
get_all_resources(package_contains = NULL, resource_contains = NULL)
}
\arguments{
\item{package_contains}{a character string containing an expression to be
used as search criteria against the packages 'title' field.}

\item{resource_contains}{a character string containing a regular expression
to be matched against available resource names. If a character vector >
length 1 is supplied, the first element is used.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing details of all available packages and
resources, or those containing the string specified in the
\code{package_contains} and \code{resource_contains} arguments.
}
\description{
Provides an overview of all resources available from
<opendata.nhs.scot>, with the option to limit results based on both package
and resource names. The returned data.frame can be used to look-up package
and resource ids and is useful for exploring the available data sets.
}
\examples{

get_all_resources()
get_all_resources(package_contains = "standard-populations")
get_all_resources(
  package_contains = "standard-populations", resource_contains = "European"
)

}
