% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_additional_info.R
\name{get_datasets_additional_info}
\alias{get_datasets_additional_info}
\title{returns a list of dataset names alongwith additional information,
such as the amount of resources and the date it was last updated}
\usage{
get_datasets_additional_info(dataset_name)
}
\arguments{
\item{dataset_name}{}
}
\description{
returns a list of dataset names alongwith additional information,
such as the amount of resources and the date it was last updated
}
