% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_resource_id.R
\name{get_latest_resource_id}
\alias{get_latest_resource_id}
\title{Get the latest Open Data resource ID associated
with a dataset}
\usage{
get_latest_resource_id(dataset_name)
}
\arguments{
\item{dataset_name}{name of the dataset as found on
\href{https://www.opendata.nhs.scot/}{NHS Open Data platform}}
}
\value{
a string of the resource ID that was uploaded to
the dataset most recently.
}
\description{
Returns the ID of the resource
that was uploaded to a dataset most recently.
This ID can be used as a parameter within
\code{\link[=get_resource]{get_resource()}}
}
