% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_all_resources.R
\name{list_all_resources}
\alias{list_all_resources}
\title{Deprecated: please use \code{list_resources()} instead.}
\usage{
list_all_resources(dataset_contains = NULL, resource_contains = NULL)
}
\arguments{
\item{dataset_contains}{A character string containing an expression to be
used as search criteria against the dataset 'title' field.}

\item{resource_contains}{A character string containing a regular expression
to be matched against available resource names.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing details of all available datasets and
resources, or those containing the string specified in the
\code{dataset_contains} and \code{resource_contains} arguments.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}

\code{list_all_resources()} was renamed to \code{\link[=list_resources]{list_resources()}} for naming consistency.
}
\examples{

list_resources()
list_resources(dataset_contains = "standard-populations")
list_resources(
  dataset_contains = "standard-populations", resource_contains = "European"
)

}
\keyword{internal}
