% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{Lists all available datasets}
\usage{
list_datasets()
}
\value{
A tibble.
}
\description{
\code{list_datasets()} shows all of the datasets hosted on the PHS Open Data
Platform.
}
\section{Lifecycle}{


\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#superseded}{\figure{lifecycle-superseded.svg}{options: alt='[Superseded]'}}}{\strong{[Superseded]}}

\code{list_datasets()} has been superseded by \code{list_resources()}.
While \code{list_datasets()} only returns a list of dataset packages,
\code{list_resources()} provides a more comprehensive and flexible
interface for exploring the PHS Open Data platform. It returns both
datasets and their associated resources in a single tibble, and
supports filtering by dataset titles or resource names.
}

\seealso{
\code{\link[=list_resources]{list_resources()}}
}
