% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{package_show_url}
\alias{package_show_url}
\title{Creates the URL for the package_show end-point}
\usage{
package_show_url()
}
\value{
a url
}
\description{
Creates the URL for the package_show end-point
}
