% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource.R
\name{parse_col_select}
\alias{parse_col_select}
\title{Create fields query for GET}
\usage{
parse_col_select(col_select)
}
\arguments{
\item{col_select}{a character vector identifying the columns to select.}
}
\value{
a character string
}
\description{
Produces a comma separated list
}
