% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse_error.R
\name{parse_error}
\alias{parse_error}
\title{Generates error message from an "error" element of an \code{httr::content} object}
\usage{
parse_error(error)
}
\arguments{
\item{error}{"error" element of an object produced by \code{httr::content}}
}
\value{
character string
}
\description{
Generates error message from an "error" element of an \code{httr::content} object
}
