% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource.R
\name{parse_row_filters}
\alias{parse_row_filters}
\title{Create json 'dict' from named list or vector}
\usage{
parse_row_filters(row_filters)
}
\value{
a json as a character string
}
\description{
Formats a list or named vector into a valid query
}
