% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/phs_GET.R
\name{phs_GET}
\alias{phs_GET}
\title{Send a GET request to the PHS CKAN API}
\usage{
phs_GET(action, query, verbose = FALSE)
}
\arguments{
\item{action}{The API endpoint you want to use, e.g., "package_show" / "datastore_search".}

\item{query}{The query to pass to the endpoint defined by the action argument.}

\item{verbose}{TRUE or FALSE. If TRUE, a success message will be printed to the console.}
}
\value{
content of a httr::GET request
}
\description{
Send a GET request to the PHS CKAN API
}
