% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/request_url.R
\name{request_url}
\alias{request_url}
\title{Produces a URL for a GET request to opendata.nhs.scot}
\usage{
request_url(action, query)
}
\arguments{
\item{action}{The API endpoint you want to use, e.g., "package_show" / "datastore_search".}

\item{query}{The query to pass to the endpoint defined by the action argument.}
}
\value{
a URL as a character string
}
\description{
Produces a URL for a GET request to opendata.nhs.scot
}
