% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_resources.R
\name{show_resources}
\alias{show_resources}
\title{Show all available resources for a dataset}
\usage{
show_resources(dataset_name)
}
\arguments{
\item{dataset_name}{}
}
\value{
a \link[tibble:tibble-package]{tibble} with the data
}
\description{
Show all available resources for a dataset
}
\examples{
show_resources("weekly-accident-and-emergency-activity-and-waiting-times")
}
