% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_dump_check.R
\name{use_dump_check}
\alias{use_dump_check}
\title{Determines if the datastore dump should be used}
\usage{
use_dump_check(query, rows)
}
\arguments{
\item{query}{a list of items to query}

\item{rows}{the number of rows user is requesting}
}
\value{
a logical value. TRUE indicates that the dump should be used
}
\description{
Determines if the datastore dump should be used
}
