% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset.R
\name{check_dataset_name}
\alias{check_dataset_name}
\title{Check if a dataset name is valid}
\usage{
check_dataset_name(dataset_name)
}
\arguments{
\item{dataset_name}{a resource ID}
}
\value{
TRUE / FALSE indicating the validity of the dataset name
}
\description{
Used to attempt to validate a dataset name before submitting it to the API
}
