% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource.R
\name{check_res_id}
\alias{check_res_id}
\title{Check if a resource ID is valid}
\usage{
check_res_id(res_id)
}
\arguments{
\item{res_id}{a resource ID}
}
\value{
TRUE / FALSE indicating the validity of the res_id
}
\description{
Used to attempt to validate a res_id before submitting it to the API
}
