% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_additional_info.R
\name{get_dataset_additional_info}
\alias{get_dataset_additional_info}
\title{get a datasets additional info}
\usage{
get_dataset_additional_info(dataset_name)
}
\arguments{
\item{dataset_name}{name of the dataset as found on
\href{https://www.opendata.nhs.scot/}{NHS Open Data platform}}
}
\value{
a \link[tibble:tibble-package]{tibble} with the data
}
\description{
\code{get_dataset_additional_info()} returns a list of dataset names along with
additional information, such as the amount of resources and the date it was
last updated.
}
\examples{
get_dataset("gp-practice-populations")
}
