% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_additional_info.R
\name{get_dataset_additional_info}
\alias{get_dataset_additional_info}
\title{get a datasets additional info}
\usage{
get_dataset_additional_info(dataset_name)
}
\arguments{
\item{dataset_name}{Name of the dataset as found on \href{https://www.opendata.nhs.scot/}{NHS Open Data platform} (character).}
}
\value{
a \link[tibble:tibble-package]{tibble} with the data
}
\description{
\code{get_dataset_additional_info()} returns a tibble of dataset names along with
the amount of resources it has and the date it was last updated.Last updated
is taken to mean the most recent date a resource within the dataset was
created or modified.
}
\examples{
get_dataset_additional_info("gp-practice-populations")
}
