% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_resource.R
\name{get_latest_resource}
\alias{get_latest_resource}
\title{Get the latest resource from a dataset}
\usage{
get_latest_resource(
  dataset_name,
  rows = NULL,
  row_filters = NULL,
  col_select = NULL,
  include_context = TRUE
)
}
\arguments{
\item{dataset_name}{Name of the dataset as found on the
\href{https://www.opendata.nhs.scot}{NHS Open Data platform} (character).}

\item{rows}{(optional) Maximum number of rows to return (integer).}

\item{row_filters}{(optional) A named list or vector specifying values of
columns/fields to keep (e.g., list(Date = 20220216, Sex = "Female")).}

\item{col_select}{(optional) A character vector containing the names of
desired columns/fields (e.g., c("Date", "Sex")).}

\item{include_context}{(optional) If \code{TRUE}, additional information about the
resource will be added as columns to the data. Defaults to \code{TRUE} for this
function.}
}
\value{
A \link[tibble:tibble-package]{tibble} with the data.
}
\description{
The Scottish Health and Social Care Open Data platform
hosts some datasets that are updated regularly with new resources.
For some of these datasets, the latest resource is the one that
is most likely to be of interest. This function attempts to
identify the latest resource for a given dataset and returns
it as a tibble.
}
\seealso{
\code{\link[=get_resource]{get_resource()}} for downloading a single resource from a dataset.

\code{\link[=get_dataset]{get_dataset()}} for downloading all resources from a dataset.
}
