% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_latest_resource_id.R
\name{get_latest_resource_id}
\alias{get_latest_resource_id}
\title{get_latest_resource_id}
\usage{
get_latest_resource_id(dataset_name)
}
\arguments{
\item{dataset_name}{name of the dataset as found on
\href{https://www.opendata.nhs.scot/}{NHS Open Data platform}}
}
\value{
a string with the resource id
}
\description{
to be confident that the resource returned is the one intended
two conditions have to be met. It has to appear at the top of
of the resource list as shown on the open data platform.
The order they are returned via the api is the same
as they appear on the open data platform. It also
has to have the most recent date created
}
\details{
There are only some datasets that this functionality
is relevant to, these are listed within applicable
datasets and are the datasets that keep historic
resources instead of over writing them.
}
