% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_resource.R
\name{get_resource}
\alias{get_resource}
\title{Get Open Data resource}
\usage{
get_resource(res_id, rows = NULL)
}
\arguments{
\item{res_id}{The resource ID as found on
\href{https://www.opendata.nhs.scot/}{NHS Open Data platform}}

\item{rows}{(optional) specify the max number of rows to return
use this when testing code to reduce the size of the request
it will default to all data}
}
\value{
a \link[tibble:tibble-package]{tibble} with the data
}
\description{
Get Open Data resource
}
\examples{
get_resource(res_id = "a794d603-95ab-4309-8c92-b48970478c14")
}
\seealso{
\code{\link[=get_dataset]{get_dataset()}} for downloading all resources
from a given dataset.
}
