% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{Lists all available datasets}
\usage{
list_datasets(include_additional_info = F)
}
\arguments{
\item{include_additional_info}{(optional) set to true to return the number
of resources per dataset and the date the dataset was last updated}
}
\value{
A character vector.
}
\description{
\code{list_datasets()} shows all of the datasets hosted on the phs open data platform.
}
\examples{
head(list_datasets())
head(list_datasets(include_additional_info = T))

}
