% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_datasets.R
\name{list_datasets}
\alias{list_datasets}
\title{Lists all available datasets}
\usage{
list_datasets()
}
\value{
A tibble.
}
\description{
\code{list_datasets()} shows all of the datasets hosted on the PHS Open Data
Platform.
}
\examples{
\dontshow{if (isTRUE(length(curl::nslookup("www.opendata.nhs.scot", error = FALSE)) > 0L)) withAutoprint(\{ # examplesIf}
head(list_datasets())
\dontshow{\}) # examplesIf}
}
