% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_resources.R
\name{list_resources}
\alias{list_resources}
\title{Provides an overview of all resources available on the Open Data platform.}
\usage{
list_resources(
  dataset_contains = NULL,
  resource_contains = NULL,
  dataset_name = lifecycle::deprecated()
)
}
\arguments{
\item{dataset_contains}{A character string containing an expression to be
used as search criteria against the dataset 'title' field.}

\item{resource_contains}{A character string containing a \href{https://www.geeksforgeeks.org/dsa/write-regular-expressions/}{regular expression}
to be matched against available resource names.}

\item{dataset_name}{Deprecated. Use dataset_contains instead.}
}
\value{
A \link[tibble:tibble-package]{tibble} containing details of all available datasets and
resources, or those containing the string specified in the
\code{dataset_contains} and \code{resource_contains} arguments.
}
\description{
Provides an overview of all resources available from
\href{https://www.opendata.nhs.scot/}{opendata.nhs.scot}, with the option to limit results based on both dataset
and resource names. The returned tibble can be used to look-up dataset
and resource ids, and is useful for exploring the available data sets.
}
\examples{

list_resources()
list_resources(dataset_contains = "standard-populations")
list_resources(
  dataset_contains = "standard-populations", resource_contains = "European"
)

}
