% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/show_datasets.R
\name{show_datasets}
\alias{show_datasets}
\title{Show all available datasets}
\usage{
show_datasets()
}
\value{
A character vector.
}
\description{
\code{show_datasets()}Shows all datasets available form phs open data platform.
}
\examples{
show_datasets()

}
