% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/suggest_dataset_name.R
\name{suggest_dataset_name}
\alias{suggest_dataset_name}
\title{Throw an error with suggested dataset name, if possible.}
\usage{
suggest_dataset_name(dataset_name)
}
\arguments{
\item{dataset_name}{Name of the dataset as found on \href{https://www.opendata.nhs.scot/}{NHS Open Data platform} (character).}
}
\description{
Throw an error with suggested dataset name, if possible.
}
